# Conflicts:
#	src/Processors/Transforms/BufferedShardByHashTransform.cpp

+++ src/Processors/Transforms/ScatterByPartitionTransform.cpp
@@ -143,7 +143,10 @@ void ScatterByPartitionTransform::generateOutputChunks()
 
     chassert(!columns.empty());
 
-    hash.assign(num_rows, WEAK_HASH32_INITIAL_VALUE);
+    /// Cast to size_t to select the (count, value) overload of `assign`: on Darwin `UInt64` is
+    /// `unsigned long long` while `size_t` is `unsigned long`, so without the cast the iterator-pair
+    /// overload would be deduced and fail to compile.
+    hash.assign(static_cast<size_t>(num_rows), WEAK_HASH32_INITIAL_VALUE);
 
     for (const auto & column_number : key_columns)
         columns[column_number]->computeHashInto(0, num_rows, hash.data(), false);
